+++ src/bluezdbus/BluezAdapter.cpp
@@ -8,7 +8,7 @@ BluezAdapter::BluezAdapter(SimpleDBus::Connection* conn, std::string path, Simpl
     : _conn(conn), _path(path), Adapter1{conn, path} {
     // std::cout << "Creating BluezAdapter" << std::endl;
 
-    Properties::PropertiesChanged = [&](std::string interface, SimpleDBus::Holder changed_properties,
+    PropertyHandler::PropertiesChanged = [&](std::string interface, SimpleDBus::Holder changed_properties,
                                         SimpleDBus::Holder invalidated_properties) {
         if (interface == "org.bluez.Adapter1") {
             Adapter1::set_options(changed_properties, invalidated_properties);
@@ -28,7 +28,7 @@ BluezAdapter::~BluezAdapter() {
 
 bool BluezAdapter::process_received_signal(SimpleDBus::Message& message) {
     if (message.get_path() == _path) {
-        if (Properties::process_received_signal(message)) return true;
+        if (PropertyHandler::process_received_signal(message)) return true;
     } else {
         for (auto& [device_path, device] : _devices) {
             if (device->process_received_signal(message)) {