+++ src/bluezdbus/interfaces/GattCharacteristic1.h
@@ -5,7 +5,7 @@
 #include <string>
 #include <vector>
 
-class GattCharacteristic1 : public SimpleDBus::Interfaces::PropertyHandler {
+class GattCharacteristic1 : public SimpleDBus::Interfaces::PropertyHandler, public SimpleDBus::Properties {
   private:
     static const std::string _interface_name;