+++ src/bluezdbus/interfaces/GattService1.cpp
@@ -2,7 +2,7 @@
 
 #include <iostream>
 
-GattService1::GattService1(SimpleDBus::Connection* conn, std::string path) : _conn(conn), _path(path) {
+GattService1::GattService1(SimpleDBus::Connection* conn, std::string path) : _conn(conn), _path(path), Properties{conn, "org.bluez", path}, PropertyHandler(path) {
     // std::cout << "Creating org.bluez.GattService1: " << path << std::endl;
 }